refactor(docs): route docs collection through content-processor
